changed to write stats rec when recorded
